user_uploads/892054522032160791/bot_1/requirements.txt
discord.py==2.3.0
python-dotenv==1.0.0